\name{RcppTBASS-package}
\alias{RcppTBASS-package}
\alias{RcppTBASS}
\docType{package}
\title{
\packageTitle{RcppTBASS}
}
\description{
\packageDescription{RcppTBASS}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{RcppTBASS}
\packageIndices{RcppTBASS}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{RcppTBASS}

Maintainer: \packageMaintainer{RcppTBASS}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
